shiftfs: fix buggy unlink logic
The way we messed with setting i_nlink was brittle and wrong. We also
missed a drop_nlink() which is crucial since i_nlink affects whether a
dentry is cleaned up on dput(). With this I cannot reproduce the bug
anymore where shiftfs misleads zfs into believing that a deleted file
can not be removed from disk because it is still referenced.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>